refactor: move modules to templates library
