+++ src/wazuh_testing/constants/templates/syscollector.py
